Add min_individuals column
When choosing a minimal set of mutation operators, we want to minimise the number of mutants as well, not just the number of operators. For instance, it could be better to have more operators that are very specific, if it means having fewer individuals to go through.

This commit adds a min_individuals column with the minimum number of mutants across the subsuming operators that each operator was part of. The results so far indicate that the 10 operators picked when sampling mutants with duplicate rows would have fewer mutants in total (40) compared to the 49 from the 13 operators picked when discarding mutants with duplicate rows.